// src/redux/game-reducer/interfaces.ts
export interface IDataState {
    numArray: any;
    btnValues: any;
    btnText: any;
    data: any;
    backgroundColors: any;
    textColors: any;
    outerBorderColor: any;
    outerBorderWidth: any;
    innerBorderColor: any;
    innerBorderWidth: any;
    innerRadius: any
    radiusLineColor: any;
    radiusLineWidth: any;
    fontSize: any;
    textDistance: any;
    width: any;
    height: any;
    perpendicularText: any;
    curNum: any;
}

export interface SetArray {
    type: string;
    numArray: any;
}
export interface SetNum {
    type: string;
    curNum: any;
}